fix(tests): flaky test case in BLMPOP command
The root cause should be the blocking callback was triggered after
pushing the key1's elements. And the key1 will be consumed before key2
because BLMPOP command will check keys by order when waking up. We can
fix by putting the key1 before key2, so that the key order is always
correct even if the blocking callback was delayed.